alpr_service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2
paddlepaddle==2.5.1
paddleocr==2.7.0.3
cloudinary==1.36.0
python-dotenv==1.0.0 